requirements.txt
Django>=5.0,<5.1
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery
redis
yfinance
google-genai
django-environ
django-cors-headers
pillow